+++ src/MeritooCommonBundle.php
@@ -1,0 +1,24 @@
+<?php
+
+/**
+ * (c) Meritoo.pl, http://www.meritoo.pl
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Meritoo\CommonBundle;
+
+use Symfony\Component\HttpKernel\Bundle\Bundle;
+
+/**
+ * The Meritoo Common Bundle.
+ * Main class.
+ *
+ * @author    Meritoo <github@meritoo.pl>
+ * @copyright Meritoo <http://www.meritoo.pl>
+ */
+class MeritooCommonBundle extends Bundle
+{
+}